# apilayer/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
...
    Inventory:
        properties:
            barcode:
                type: string
                x-go-name: Barcode
            bought_at:
                type: string
                x-go-name: BoughtAt
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
            created_by:
                type: string
                x-go-name: CreatedBy
            creator_name:
                type: string
                x-go-name: CreatorName
            description:
                type: string
                x-go-name: Description
            id:
                type: string
                x-go-name: ID
            is_returnable:
                type: boolean
                x-go-name: IsReturnable
            location:
                type: string
                x-go-name: Location
            max_height:
                type: string
                x-go-name: MaxHeight
            max_weight:
                type: string
                x-go-name: MaxWeight
            min_height:
                type: string
                x-go-name: MinHeight
            min_weight:
                type: string
                x-go-name: MinWeight
            name:
                type: string
                x-go-name: Name
            price:
                format: double
                type: number
                x-go-name: Price
            quantity:
                format: int64
                type: integer
                x-go-name: Quantity
            return_datetime:
                format: date-time
                type: string
                x-go-name: ReturnDateTime
            return_location:
                type: string
                x-go-name: ReturnLocation
            return_notes:
                type: string
                x-go-name: ReturnNotes
            sku:
                type: string
                x-go-name: SKU
            status:
                type: string
                x-go-name: Status
            storage_location_id:
                type: string
                x-go-name: StorageLocationID
            updated_at:
                format: date-time
                type: string
                x-go-name: UpdatedAt
            updated_by:
                type: string
                x-go-name: UpdatedBy
            updater_name:
                type: string
                x-go-name: UpdaterName
        title: Inventory ...
        type: object
        x-go-package: github.com/mohit2530/communityCare/model
...
    MessageResponse:
        properties:
            Message:
                type: string
        title: MessageResponse ...
        type: object
        x-go-package: github.com/mohit2530/communityCare
...
info:
    description: |-
        # Documentation for the Asset alert api layer.

        Scehmes: http
    title: alert Api Layer
    version: 1.0.0
paths:
    /api/profile/{id}/inventories:
        post:
            operationId: removeSelectedInventory
            parameters:
                - description: The id of the selected user
                  in: path
                  name: id
                  required: true
                  type: string
                - description: The inventory object to remove from the db
                  in: body
                  name: Inventory
                  required: true
                  schema:
                    $ref: '#/definitions/object'
                    description: The inventory object to remove from the db
                    type: object
            responses:
                "200":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: '# Update selected inventory with details.'
            tags:
                - RemoveSelectedInventory
        put:
            operationId: updateSelectedInventory
            parameters:
                - description: The id of the selected user
                  in: path
                  name: id
                  required: true
                  type: string
                - description: The inventory object to add into the db
                  in: body
                  name: Inventory
                  required: true
                  schema:
                    $ref: '#/definitions/object'
                    description: The inventory object to add into the db
                    type: object
            responses:
                "200":
                    description: Inventory
                    schema:
                        $ref: '#/definitions/Inventory'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: '# Update selected inventory with details.'
            tags:
                - UpdateSelectedInventory
    /api/profile/{id}/inventories/bulk:
        post:
            operationId: addInventoryInBulk
            parameters:
                - description: The id of the selected inventory
                  in: path
                  name: id
                  required: true
                  type: string
                - description: The list of inventories to add into the db to support bulk upload
                  in: body
                  name: InventoryListRequest
                  required: true
                  schema:
                    $ref: '#/definitions/object'
                    description: The list of inventories to add into the db to support bulk upload
                    type: object
            responses:
                "200":
                    $ref: '#/responses/model.Inventory'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: '# Bulk upload inventory list. returns the list of uploaded items.'
            tags:
                - AddInventoryInBulk
...
    /api/v1/profile/{id}/inventories:
        get:
            operationId: getAllInventories
            parameters:
                - description: The userID of the selected user
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "200":
                    description: Inventory
                    schema:
                        items:
                            $ref: '#/definitions/Inventory'
                        type: array
                "400":
                    $ref: '#/responses/Message'
                "404":
                    $ref: '#/responses/Message'
                "500":
                    $ref: '#/responses/Message'
            summary: Retrieves the list of assets that belong to the selected user
            tags:
                - GetAllInventories
    /api/v1/profile/{id}/inventories/{assetID}:
        get:
            operationId: updateAssetColumn
            parameters:
                - description: The userID of the selected user
                  in: path
                  name: id
                  required: true
                  type: string
                - description: The asssetID of the asset details.
                  in: path
                  name: invID
                  required: true
                  type: string
                - description: The object containing the details to update the selected asset with.
                  in: body
                  name: UpdateAssetColumn
                  required: true
                  schema:
                    $ref: '#/definitions/UpdateAssetColumn'
                    description: The object containing the details to update the selected asset with.
                    type: object
            responses:
                "200":
                    description: Inventory
                    schema:
                        $ref: '#/definitions/Inventory'
                "400":
                    $ref: '#/responses/Message'
                "404":
                    $ref: '#/responses/Message'
                "500":
                    $ref: '#/responses/Message'
            summary: '# Updates the selected quantity or price column field with passed in data field. All colaborators can update selected field.'
            tags:
                - UpdateAssetColumn
    /api/v1/profile/{id}/inventories/{invID}:
        get:
            operationId: getInventoryByID
            parameters:
                - description: The userID of the selected user
                  in: path
                  name: id
                  required: true
                  type: string
                - description: The inventoryID of the inventory details.
                  in: path
                  name: invID
                  required: true
                  type: string
            responses:
                "200":
                    description: Inventory
                    schema:
                        $ref: '#/definitions/Inventory'
                "400":
                    $ref: '#/responses/Message'
                "404":
                    $ref: '#/responses/Message'
                "500":
                    $ref: '#/responses/Message'
            summary: '# Retrieves the selected inventory that matches the passed in ID created by the user.'
            tags:
                - GetInventoryByID
...
produces:
    - application/json
swagger: "2.0"